requirement.txt
altgraph                  0.17.3
certifi                   2022.12.7
libusb1                   3.0.0
ndeflib                   0.3.3
nfcpy                     1.0.4
numpy                     1.24.2
pandas                    2.0.0
pefile                    2023.2.7
Pillow                    9.5.0
pip                       23.0.1
plyer                     2.1.0
pyDes                     2.0.1
pyinstaller               5.10.0
pyinstaller-hooks-contrib 2023.2
pyserial                  3.5
pystray                   0.19.4
python-dateutil           2.8.2
pytz                      2023.3
pywin32-ctypes            0.2.0
setuptools                65.6.3
six                       1.16.0
tzdata                    2023.3
wheel                     0.38.4
wincertstore              0.2